requirements.txt
flask
flask-sqlalchemy
flask-security
email-validator
mysqlclient
flask-marshmallow
bcrypt
